Enable too-small-loop-variable check (#15984)
### Ticket
#15123

### Problem description
Loop variables should be capable of covering the range of values that
are will be compared against. More info:
https://clang.llvm.org/extra/clang-tidy/checks/bugprone/too-small-loop-variable.html

### What's changed
The issues fixed here should not have ever exhibited as bugs. However
now that we're clean we can enforce the check globally which can save us
from future bugs.

### Checklist
- [x] Post commit CI passes
https://github.com/tenstorrent/tt-metal/actions/runs/12307379048

+++ tt_metal/impl/program/program.cpp
@@ -1430,7 +1430,7 @@ const std::vector<SubDeviceId> &detail::Program_::determine_sub_device_ids(const
                 uint32_t num_intersections = 0;
                 uint32_t num_cores = 0;
                 for (const auto& kg : program_kgs) {
-                    for (uint8_t i = 0; i < device->num_sub_devices(); ++i) {
+                    for (size_t i = 0; i < device->num_sub_devices(); ++i) {
                         const auto& sub_device_cores = device->worker_cores(core_type, SubDeviceId{i});
                         auto intersection = sub_device_cores.intersection(kg.core_ranges);
                         if (intersection.size() > 0) {
